Added the form with php code to submit.php
So I've been going through tutorials and I think this works (hopefully).
I added the php code that should be able to add the necessary info
collected from the form to the Idea table in the database. Since I don't
have the info for connecting to the database, I didn't add the database
connection lines. I really hope this works!