Added the web crawler
The web crawler retrieves descriptions of the gathered apps in the app
store. It also gathers all kinds of other info but we can easily get rid
of those. Make sure you read the readme file and set the operation type
in the code to "store" if you want to retrieve apps. This folder has the
csv files that got created from retrieving apps and their info from when
I started the crawler and if you run it on your system, it will write
over those files. If you want to keep them, then just copy all the files
to another folder and delete all of the csv files. Then the crawler will
create new csv files with the apps that you retrieve.