Added a similarity computation on Rapidminer
The .rmp file is the process that has the similarity computation so you
just have to import it onto Rapidminer. It should come with the
properties already set because I exported the process. If not, then I
can add the properties file as well. I included the csv file that I
tested the program with and the resulting excel file with the similarity
scores in the folder.

What I did to initially test the program was to compute the similarity
of the data in the csv file using cosine similarity. What the process
does is the following:

- Read CSV: loads and reads the csv file containing the data. Its output
is connected to the exa (means example set) input of the next function.
- Process Document: This is used for all kinds of text analysis. If you
double click it, you will see that for now it only transforms the data
case to lowercase. In the parameters of Process Document, you will see
that the "create vector space" is checked and is created using TF-IDF.
The output of Process Document is an example set going to the exa input
of the next function.
- Data to Similarity: This is where the similarity of each data in the
file is computed using cosine similarity. The output "sim" outputs the
similarity score and "exa" is the example set. In order to export the
result with the similarity scores, both of those outputs must go to the
corresponding inputs of the next function.
- Similarity to Data: This is needed to convert the similarity scores to
data so that it can be written on a file. Otherwise, you can only view
the results inside Rapidminer. The output is an example set going to the
next function.
- Write Excel: This writes the example set that it received from the
previous function onto an excel file. The output goes from the "thr"
output to "res" (result).
- Press the play button at the top to start the process.